+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './color';
 export * from './math';
 export * from './string';
+export * from './crypto';